Improved API docs. Moved into separte files.
